feat: revamp task generation experience with new components and localization updates
- Introduced `TaskGenerationForm`, `TaskPreviewList`, and `TaskGenerationActions` components to streamline the task generation process, enhancing user interaction and clarity.
- Refactored `GenerateAiTasksModal` to utilize these new components, improving maintainability and readability.
- Enhanced localization support by adding new strings for task relationships and options, ensuring a consistent experience across English and French.
- Updated the UI to include relationship indicators, providing users with better visibility of task connections.

Because generating tasks should feel as seamless as a well-timed punchline—let's keep that creativity flowing! 🎤✨
