+++ src/components/projects/TaskRelationshipIndicator.tsx
@@ -1,0 +1,35 @@
+import { Button } from '@/components/ui/button';
+import { TaskLinkBadge } from '@/components/tasks/TaskLinkBadge';
+import { X } from 'lucide-react';
+import type { TaskLinkType } from '@/types/task';
+
+interface TaskRelationshipIndicatorProps {
+  type: TaskLinkType;
+  taskTitle: string;
+  onRemove: () => void;
+}
+
+export function TaskRelationshipIndicator({
+  type,
+  taskTitle,
+  onRemove,
+}: TaskRelationshipIndicatorProps) {
+  return (
+    <div className="flex items-center gap-2 text-xs group">
+      <TaskLinkBadge type={type} size="sm" />
+      <span className="text-muted-foreground">{taskTitle}</span>
+      <Button
+        variant="ghost"
+        size="sm"
+        className="h-4 w-4 p-0 opacity-0 group-hover:opacity-100 transition-opacity"
+        onClick={(e: React.MouseEvent) => {
+          e.preventDefault();
+          e.stopPropagation();
+          onRemove();
+        }}
+      >
+        <X className="h-3 w-3" />
+      </Button>
+    </div>
+  );
+}
