+++ src/locales/fr.json
@@ -825,7 +825,8 @@
       "backToPrompt": "Retour au prompt",
       "regenerate": "Régénérer",
       "addSelected": "Ajouter les tâches sélectionnées",
-      "relationships": "{{count}} relation | {{count}} relations"
+      "relationships_one": "{{count}} relation",
+      "relationships_other": "{{count}} relations"
     }
   }
 }
